+++ test/beckon_signal_child.clj
@@ -1,0 +1,10 @@
+(ns beckon-signal-child
+  (:gen-class)
+  (:require [beckon :as beckon]))
+
+(defn -main [signal]
+  (reset! (beckon/signal-atom signal)
+          [(fn [] (println "HANDLED") (flush))])
+  (println "READY")
+  (flush)
+  @(promise))